requirements.txt
py-cord==2.3.2
python-dotenv==0.21.0
psutil==5.9.4
inflect==6.0.2
